+++ src/main/java/com/mycompany/app/week3/code2/InterfacesExercise21.java
@@ -1,0 +1,5 @@
+package com.mycompany.app.week3.code2;
+
+public class InterfacesExercise21 {
+    
+}
